only-root

+++ snapshot.sh
@@ -38,13 +38,10 @@ if [ "make" == "$1" ]; then
 
     done
 
-    #reset-fstab
-    if [ "${rootsnapshot}" == "y" ]; then
-        cp /etc/fstab.example /etc/fstab
-    fi
-
     #stable-snapshot-boot
-    if [ -f "/boot/arch-uefi.conf.example" ]; then
+    if [ -f "/boot/arch-uefi.conf.example" ] && [ "${rootsnapshot}" == "y" ]; then
+        #reset-fstab
+        cp /etc/fstab.example /etc/fstab
 
         cp "$(echo $(find /boot/ -name "initramfs*.img") | cut -d" " -f2)" /boot/initramfs-linux-stable.img
         cp "$(echo $(find /boot/ -name "vmlinuz*") | cut -d" " -f1)" /boot/vmlinuz-stable
